+++ src/main/java/net/silentchaos512/gems/lib/chaosbuff/CostConditions.java
@@ -2,6 +2,7 @@ package net.silentchaos512.gems.lib.chaosbuff;
 
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.player.PlayerEntity;
+import net.silentchaos512.gems.init.GemsEffects;
 
 import javax.annotation.Nullable;
 import java.util.function.Predicate;
@@ -9,6 +10,8 @@ import java.util.function.Predicate;
 public enum CostConditions {
     NO_CONDITION(p -> true),
     BURNING(Entity::isBurning),
+    FREEZING(p -> p.getActivePotionEffect(GemsEffects.freezing) != null),
+    SHOCKING(p -> p.getActivePotionEffect(GemsEffects.shocking) != null),
     FLYING(p -> p.abilities.isFlying),
     HURT(p -> p.getHealth() < p.getMaxHealth() - 0.5f),
     IN_AIR(p -> !p.onGround),