+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [3.4.2] - 2019-09-21
 ### Added
 - Wisp essence items, dropped by wisps (expect these to be worked into some existing recipes)
 - Chaos runes for insulated and grounded effects